refactor context and backend
