+++ crates/maple_renderer/src/core/mod.rs
@@ -1,4 +1,4 @@
-pub mod backend;
+pub mod context;
 pub mod buffer;
 pub mod descriptor_set;
 pub mod frame_builder;
@@ -8,6 +8,7 @@ pub mod shader;
 pub mod texture;
 
 pub use buffer::*;
+pub use context::RenderContext;
 pub use descriptor_set::*;
 pub use frame_builder::*;
 pub use pipeline::*;
