+++ examples/tutorial/main.rs
@@ -1,19 +1,9 @@
-use maple::{Engine, config::EngineConfig};
-use std::{default::Default, error::Error};
+use maple::prelude::*;
 
 // create and import the main scene module
 pub mod main_scene;
 use main_scene::MainScene;
 
-fn main() -> Result<(), Box<dyn Error>> {
-    let mut engine = Engine::init(EngineConfig {
-        window_title: "Hello, Window!".to_string(),
-        window_mode: maple::utils::config::WindowMode::Windowed,
-        ..Default::default()
-    })?;
-
-    // load the scene into the engine
-    engine.load_scene(MainScene::build());
-
-    engine.begin()
+fn main() {
+    App::new(Confi::default()).run()
 }
